前端页面 第6节：对接联调积分流程接口 定位bug中

+++ big-market-infrastructure/src/main/java/com/origamii/infrastructure/persistent/redis/IRedisService.java
@@ -69,7 +69,7 @@ public interface IRedisService {
      * @param key   key 键
      * @param value value 值
      */
-    void setAtomicLong(String key, Integer value);
+    void setAtomicLong(String key, long value);
 
     /**
      * 获取值
